add re-used types to schema file
