+++ src/component-library/Map/utils/schema.ts
@@ -1,0 +1,16 @@
+export type StyleOption = {
+  label: string;
+  uri: string;
+  image: string;
+};
+
+export type ClassIcon = {
+  backgroundColor: string;
+  color: string;
+  classUri: string;
+  iconFallbackText: string;
+  alt: string;
+  faIcon?: string | undefined;
+  faUnicode?: string | undefined;
+  shape?: string | undefined;
+};
